Add separate getUser util method for setting user-specific metadata
Add bugsnag prefix to util files
