+++ app/utils/bugsnag-get-metadata.js
