+++ app/utils/bugsnag-get-user.js
@@ -1,0 +1,3 @@
+export function getUser(/* container */) {
+  return {};
+}
